+++ catkin_ws/src/arduino/src/MotorController/src/AfroESC.h
@@ -18,9 +18,9 @@ public:
 
     virtual ~AfroESC();
 
-    virtual void setSpeed(int speed);
+    virtual void setSpeed(float speed);
 
-    virtual void setDirection(int speed);
+    virtual void setDirection(float speed);
 
     virtual void brake(bool brk);
